feat(wip): save proofs in file
