+++ cmd/ethrex_l2/src/commands/prove.rs
@@ -1,5 +1,8 @@
 use clap::Args;
-use ethrex_l2::utils::test_data_io::{generate_program_input, read_chain_file, read_genesis_file};
+use ethrex_l2::utils::{
+    prover::proving_systems::ProverType,
+    test_data_io::{generate_program_input, read_chain_file, read_genesis_file},
+};
 use ethrex_prover_lib::prover::create_prover;
 
 #[derive(Args)]
@@ -30,7 +33,7 @@ impl Command {
         let chain = read_chain_file(&self.chain);
         let program_input = generate_program_input(genesis, chain, self.block_number)?;
 
-        let mut prover = create_prover(ethrex_l2::proposer::prover_server::ProverType::RISC0);
+        let mut prover = create_prover(ProverType::RISC0);
         prover.prove(program_input).expect("proving failed");
         println!(
             "Total gas consumption: {}",
