+++ crates/l2/prover/src/main.rs
@@ -1,6 +1,6 @@
-use ethrex_l2::{
-    proposer::prover_server::ProverType,
-    utils::config::{prover_client::ProverClientConfig, read_env_file},
+use ethrex_l2::utils::{
+    config::{prover_client::ProverClientConfig, read_env_file},
+    prover::proving_systems::ProverType,
 };
 use ethrex_prover_lib::init_client;
 use std::env;
